fakenet-signer: consume @sig-net/midnight abi-serde end to end
EthereumMonitor now decodes EVM return data with deserializeEvmOutput
(raw NUL-padded schema bytes in, plain decoded values out), completing the
move of both respond-path conversions to the signet library: the Midnight
respond bytes already came from serializeRespondOutput, OutputSerializer is
gone, and the Borsh path lives inline in ChainSignatureServer.
SerializableValue gains Uint8Array for decoded byte values.

Cross-repo dev wiring, pending an npm release of the new library surface:
@sig-net/midnight and @sig-net/midnight-contract resolve via yarn link: to
the sibling midnight-integration checkout, a root resolution covers the
unpublished @sig-net/midnight-serde, and the fakenet tsconfig gains noEmit
plus allowImportingTsExtensions so `yarn tsc --noEmit` typechecks the
linked raw-TS sources. Revert the link: entries to published versions
before the next npm release.
