+++ fakenet-signer/index.ts
@@ -8,7 +8,6 @@ export {
   CpiEventParser,
   EMIT_CPI_INSTRUCTION_DISCRIMINATOR,
 } from './src/events/CpiEventParser';
-export { OutputSerializer } from './src/modules/OutputSerializer';
 export * from './src/types';
 export * from './src/modules/ChainUtils';
 export { CONFIG } from './src/config/Config';
