+++ fakenet-signer/tsconfig.json
@@ -5,6 +5,8 @@
     "lib": ["ES2022"],
     "moduleResolution": "bundler",
     "declaration": true,
+    "noEmit": true,
+    "allowImportingTsExtensions": true,
     "strict": true,
     "noImplicitReturns": true,
     "noFallthroughCasesInSwitch": true,
